# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Investment App API",
    "version": "1.0.0",
    "description": "Backend API documentation for authentication, portfolio analytics, reporting, and admin/security endpoints."
  },
  "servers": [
    {
      "url": "http://localhost:4040",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/v1/admin/deposits/{id}/approve": {
      "post": {
        "summary": "Approve manual deposit and credit funds",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "admin_note": { "type": "string", "maxLength": 500 },
                  "bank_settlement_ref": { "type": "string", "minLength": 2, "maxLength": 128 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Deposit approved and credited" },
          "403": { "description": "Forbidden or rollout disabled" },
          "409": { "description": "Deposit already reviewed" }
        }
      }
    },
    "/api/v1/admin/deposits/{id}/reject": {
      "post": {
        "summary": "Reject manual deposit",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["reason"],
                "properties": {
                  "reason": { "type": "string", "minLength": 3, "maxLength": 300 },
                  "admin_note": { "type": "string", "maxLength": 500 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Deposit rejected" },
          "403": { "description": "Forbidden or rollout disabled" },
          "409": { "description": "Deposit already reviewed" }
        }
      }
    },
    "/api/v1/admin/withdrawals/{id}/approve": {
      "post": {
        "summary": "Approve manual withdrawal and debit funds",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "admin_note": { "type": "string", "maxLength": 500 },
                  "bank_settlement_ref": { "type": "string", "minLength": 2, "maxLength": 128 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Withdrawal approved and debited" },
          "403": { "description": "Forbidden or rollout disabled" },
          "409": { "description": "Withdrawal already reviewed" }
        }
      }
    },
    "/api/v1/admin/withdrawals/{id}/reject": {
      "post": {
        "summary": "Reject manual withdrawal",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["reason"],
                "properties": {
                  "reason": { "type": "string", "minLength": 3, "maxLength": 300 },
                  "admin_note": { "type": "string", "maxLength": 500 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Withdrawal rejected" },
          "403": { "description": "Forbidden or rollout disabled" },
          "409": { "description": "Withdrawal already reviewed" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}